+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 
 ### Added
 
+- You can now configure plugins on each client in `extra_plugins`. Plugins that you previously had to define your own services can be configured on the client.
 - Support for BatchClient
-- The stopwatch plugin in included by default when using profiling. 
+- The stopwatch plugin in included by default when using profiling.
 
 ### Changed